cc-wrapper: Support derivations defining unsalted environment variables
On one hand, for multiple cc-wrappers to be used together, we need some
sort of hygiene which we achieve by mangling variables. On the other
hand, various derivations define environment variables to control
cc-wrapper's behavior---something which is defeated by proper hygiene.

As a compromise, we have cc-wrapper see how it itself is being used,
and optionally take in information from the environment as appropriate.
This means we have "out hygiene" (nothing escapes to cause problems
elsewhere) but not in hygiene (the derivation cannot control
cc-wrapper). The assumption that makes this work is that while a
derivation may include multiple cc-wrappers as dependencies, they will
always be different types of dependencies.

Add `concatDelimit`, `predendDelimit`, and `appendDelimit` to assist
with this.