nvidia-x11: add disable32Bit arg to not include 32-bit nvidia
In some contexts, we don’t want to have to build the whole i686 stdenv
just to use the x86_64 nvidia driver. It’s hard to know ahead of time
what we want, so it’s best to leave this as an overridable option.

(cherry picked from commit 31d8669a72667f288591669b803f560a5d5e1756)